Fix teardown of the security.Checkers layer.
This was a fun one.
The zope.security.checker module has a module-global _checkers dict.
This layer tries to manage it as a stack, where pushing to the stack
copies the current dict and popping from the stack sets
zope.security.checker._checkers back to the copy that was stored.

But, it turns out that _checkers is imported from a C extension module,
and that module keeps a pointer to the original dict that is not
affected by the plone.testing stack pop. So subsequent lookups continue
to use the old dict.

To fix this I changed the stack pop operation to clear and update the
dict instead of replacing it, so that the pointer in the C module
is still valid.

There was a similar but different problem with the pure-Python implementation:
it sets _getChecker = _checkers.get initially and _getChecker was not updated
by the stack pop. My workaround should also deal with this case.