Fix zope.uninstallProduct for Products installed with pip.
In [PR 105](https://github.com/plone/plone.testing/pull/105) a few months ago I fixed `installProduct` for pip-installed packages in the `Products` namespace.  But a similar fix is needed in `uninstallProduct`.
It is less of a problem there, but still the cleanup seems good to run.

Also fixed the message that you get when `uninstallProduct` is called on a product that *really* cannot be found.  Previously it would confusingly report "Could not install product", so "install" instead of "uninstall".

Also, the code that was run for uninstalling non-Products packages, was not really doing anything, or at least it was not cleaning up anything.  I have fixed that.